+++ mayan/apps/mailer/tests/test_document_file_views.py
@@ -57,7 +57,7 @@ class MailDocumentFileViewTestCase(
         mail_messages = len(mail.outbox)
 
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -80,7 +80,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_link
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -97,7 +97,7 @@ class MailDocumentFileViewTestCase(
 
         self.assertEqual(events[0].action_object, self.test_document_file)
         self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self.test_user_mailer)
+        self.assertEqual(events[0].target, self._test_user_mailer)
         self.assertEqual(events[0].verb, event_email_sent.id)
 
     def test_trashed_document_file_send_link_single_view_with_full_access(self):
@@ -110,7 +110,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_link
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self.test_document.delete()
@@ -166,7 +166,7 @@ class MailDocumentFileViewTestCase(
         mail_messages = len(mail.outbox)
 
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -189,7 +189,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_link
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -206,7 +206,7 @@ class MailDocumentFileViewTestCase(
 
         self.assertEqual(events[0].action_object, self.test_document_file)
         self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self.test_user_mailer)
+        self.assertEqual(events[0].target, self._test_user_mailer)
         self.assertEqual(events[0].verb, event_email_sent.id)
 
     def test_trashed_document_file_send_link_multiple_view_with_full_access(self):
@@ -219,7 +219,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_link
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self.test_document.delete()
@@ -275,7 +275,7 @@ class MailDocumentFileViewTestCase(
         mail_messages = len(mail.outbox)
 
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -298,7 +298,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_attachment
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -316,7 +316,7 @@ class MailDocumentFileViewTestCase(
 
         self.assertEqual(events[0].action_object, self.test_document_file)
         self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self.test_user_mailer)
+        self.assertEqual(events[0].target, self._test_user_mailer)
         self.assertEqual(events[0].verb, event_email_sent.id)
 
     def test_trashed_document_file_attachment_send_single_view_with_full_access(self):
@@ -329,7 +329,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_attachment
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self.test_document.delete()
@@ -385,7 +385,7 @@ class MailDocumentFileViewTestCase(
         mail_messages = len(mail.outbox)
 
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -408,7 +408,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_attachment
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self._clear_events()
@@ -426,7 +426,7 @@ class MailDocumentFileViewTestCase(
 
         self.assertEqual(events[0].action_object, self.test_document_file)
         self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self.test_user_mailer)
+        self.assertEqual(events[0].target, self._test_user_mailer)
         self.assertEqual(events[0].verb, event_email_sent.id)
 
     def test_trashed_document_file_attachment_send_multiple_view_with_full_access(self):
@@ -439,7 +439,7 @@ class MailDocumentFileViewTestCase(
             permission=permission_send_document_file_attachment
         )
         self.grant_access(
-            obj=self.test_user_mailer, permission=permission_user_mailer_use
+            obj=self._test_user_mailer, permission=permission_user_mailer_use
         )
 
         self.test_document.delete()